Move spline sampling out of gerstner comp
