+++ crest/Assets/Crest/Crest/Scripts/Shapes/ShapeGerstnerSplineHandling.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3c8e2ffd82fc2484b80944ae10fbd697
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
